# Integer.parse/2 simpler and faster