Improves readability by being more concise ~~and avoiding meta-programming~~. It also improves performance beyond the already fast implementation from #5859.

Some numbers from executing [the benchmark](https://github.com/elixir-lang/elixir/pull/5859#issuecomment-285681896) from #5859 on my Macbook Pro (Mac OS 10.12.4, Erlang 19.3):

```
elixir 1.4.2: 8197µs
current master (3e56f6d32): 410µs
this pull-request: 149µs
```

/cc @fishcakez