@fishcakez 2017-04-28
Interesting, so the BIF is much faster at maths :D (I guess not surprising). We should be able to combine the approach in the previous PR and this approach to go even faster! Can you try using the metaprogramming approach to split the string, and then call `String.to_integer`?