@michalmuskala 2017-04-28
The idea of splitting the integer part first and using BIF to parse it is a great one and probably can be leveraged for an efficient implementation.

Unfortunately this PR changes the complexity of the algorithm from O(N) to O(N+R), where N is the length of the parseable digits and R the rest of the string. That's because it converts everything to charlist and back. This can be demonstrated by benchmarking 
```elixir
Integer.parse("1" <> String.duplicate(".", 100_000))
Integer.parse("1")
```
While current master has the same time for both calls, the proposal does not.