@josevalim 2017-04-28
@michalmuskala yup, and that can be an issue because Integer.parse can be used to take the first integer out of a potentially really large string.