@balexand 2017-04-28
I've pushed some commits. R now has no effect on the performance. The timing is slightly faster (135µs) and I've verified that the timing stays the same even with a large remainder as suggested by @michalmuskala:

I've also made the change suggested by @fishcakez. I'm passing the sign as the initial value of the accumulator. I'm using a binary for the accumulator but I could use a charlist instead. Let me know if you have feedback about this.

Thanks!