@balexand 2017-04-29
I've switched the accumulator to use a character list which gave a further performance boost from 98µs -> 67µs. The net before/after performance improvement for this pull-request is 410µs -> 67µs.