@michalmuskala 2017-04-29
What if we only counted the characters in the recursive function and then used `binary_part` to split? This way we wouldn't allocate anything.