@balexand 2017-04-29
@michalmuskala Done. It results in a slight performance improvement of 67µs -> 58µs.