@fishcakez 2017-04-30
I pushed the change, it reduces the time to parse many 10 character integer strings by 8%.